+++ arlo-client/src/components/utilities.test.ts
@@ -16,20 +16,9 @@ afterEach(() => {
 
 describe('utilities.ts', () => {
   describe('api', () => {
-    it('calls fetch with electionId', async () => {
+    it('calls fetch', async () => {
       fetchSpy.mockImplementationOnce(async () => response())
-      const result = await api('/test', { method: 'GET', electionId: '1' })
-
-      expect(result).toEqual({ success: true })
-      expect(window.fetch).toBeCalledTimes(1)
-      expect(window.fetch).toBeCalledWith('/election/1/test', {
-        method: 'GET',
-      })
-    })
-
-    it('calls fetch without electionId', async () => {
-      fetchSpy.mockImplementationOnce(async () => response())
-      const result = await api('/test', { method: 'GET', electionId: '' })
+      const result = await api('/test', { method: 'GET' })
 
       expect(result).toEqual({ success: true })
       expect(window.fetch).toBeCalledTimes(1)
@@ -40,12 +29,12 @@ describe('utilities.ts', () => {
 
     it('throws an error', async () => {
       fetchSpy.mockImplementationOnce(async () => badResponse())
-      await expect(
-        api('/test', { method: 'GET', electionId: '1' })
-      ).rejects.toThrow('A test error')
+      await expect(api('/test', { method: 'GET' })).rejects.toThrow(
+        'A test error'
+      )
       expect(window.fetch).toBeCalledTimes(1)
 
-      expect(window.fetch).toBeCalledWith('/election/1/test', {
+      expect(window.fetch).toBeCalledWith('/test', {
         method: 'GET',
       })
     })